Global refactoring of C Makefiles -> it was over-complex and didn't work with Mac. Also added support for doxygen documentation

+++ src/indexer/basefuns.c
@@ -20,19 +20,19 @@
 ##############################################################################*/
 #include <stdio.h>
 #include <stdlib.h>
-#import "base.h"
+#include "base.h"
 
 inline void *check_malloc(size_t size) {
 	void *new = malloc(size);
 	if (!new)
-		error("Error allocating %d bytes!\n", size);
+		error("Error allocating %d bytes!\n", (int)size);
 	return new;
 }
 
 inline void *check_realloc(void *ptr, size_t size) {
 	void *new = realloc(ptr, size);
 	if (!new)
-		error("Error reallocating %d bytes!\n", size);
+		error("Error reallocating %d bytes!\n", (int)size);
 	return new;
 }
